+++ src/main/java/eu/ensup/goodstudentplan/repository/IProductRepository.java
@@ -3,5 +3,5 @@ package eu.ensup.goodstudentplan.repository;
 import eu.ensup.goodstudentplan.domain.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IProductRepository extends CrudRepository<Product,Long> {
+public interface IProductRepository extends CrudRepository<Product,Integer> {
 }
